separate page components into dirs
