+++ src/components/Home/Projects.tsx
@@ -12,9 +12,9 @@ import SnapshotLogo from "public/assets/snapshot.svg";
 import UpRightArrow from "public/assets/up-right-arrow.svg";
 import { CSSProperties, useRef } from "react";
 import styled from "styled-components";
-import AnimatedLink from "./AnimatedLink";
-import { Divider } from "./Divider";
-import { BaseOuterWrapper } from "./Wrappers";
+import AnimatedLink from "../AnimatedLink";
+import { Divider } from "../Divider";
+import { BaseOuterWrapper } from "../Wrappers";
 
 type Project = {
   name: string;
