+++ src/components/pages/Osnap.tsx
@@ -1,0 +1,3 @@
+export function Osnap() {
+  
+}
